refactored title and points
