+++ src/components/KanbanBoard.tsx
@@ -6,7 +6,7 @@ import { useKanban } from "../hooks/useKanban";
 
 //Components
 import ColumnContainer from "./ColumnContainer";
-import TaskCard from "./TaskCard";
+import TaskCard from "./task/TaskCard";
 
 //UI
 import { Button } from "./@/components/ui/button";
@@ -19,7 +19,6 @@ import { ColumnProvider } from "../context/ColumnContext";
 import { TaskProvider } from "../context/TaskContext";
 
 function KanbanBoard() {
-
   const {
     columns,
     createNewColumn,
@@ -35,9 +34,8 @@ function KanbanBoard() {
     sensors,
   } = useKanban();
 
-
   return (
-    <div className="flex min-h-full w-full touch-manipulation items-start overflow-x-auto overflow-y-hidden p-8 ">
+    <div className="flex min-h-full w-full touch-manipulation items-start overflow-x-auto overflow-y-hidden p-8">
       <DndContext
         onDragStart={onDragStart}
         onDragEnd={onDragEnd}
@@ -46,10 +44,7 @@ function KanbanBoard() {
       >
         <div className="flex items-start gap-2">
           <div className="flex gap-4">
-            <SortableContext
-
-              items={columnsIds}
-            >
+            <SortableContext items={columnsIds}>
               {columns.map((col) => (
                 <ColumnProvider
                   key={col.id}
@@ -63,9 +58,7 @@ function KanbanBoard() {
           </div>
           <Button
             onClick={() => createNewColumn()}
-
             className="flex h-[80px] w-[350px] min-w-[350px] cursor-pointer items-center gap-2 rounded-lg border-2 border-b-pallette-100 bg-pallette-100 p-4 text-2xl font-semibold text-pallette-600 ring-pallette-500 hover:ring-2"
-
           >
             <FiPlusCircle />
             Add Column
@@ -79,7 +72,6 @@ function KanbanBoard() {
               duration: 300,
               easing: "cubic-bezier(0.18, 0.67, 0.6, 1.22)",
             }}
-
           >
             {activeColumn && (
               <ColumnProvider
